fix bags
пофиксил баг с открытием карточки
