+++ src/index.js
@@ -3,4 +3,3 @@ import './script/gallery';
 import './script/cardMarkup';
 import './script/onOpenCard';
 import './script/filter';
-import './script/tui-pagination';
