+++ src/partials/library.html
@@ -1,4 +1,4 @@
 <div class="container">
   <ul class="gallery card__list"></ul>
 </div>
-<div id="tui-pagination-container" class="tui-pagination"></div>
+<div id="pagination" class="tui-pagination"></div>
