+++ src/script/movieCard.js
@@ -17,15 +17,6 @@ export const movieCard = ({
   popularity = parseFloat(popularity).toFixed(1);
 
   return `
-
-  <div class="modal">
-    <button class="close__button" data-modal-close>
-        <svg class="close__icon" width="14" height="14">
-            <use class="close" href="./images/sprite/symbol-defs.svg#icon-vectorclose" width="14px" height="14px"></use>
-        </svg>
-    </button>
-    <div class="modal__container">
-
       <div class="image__place" id=${id}>
         <img class="modal-poster" src="${poster_path}" alt="${original_title}" />
       </div>
@@ -52,7 +43,5 @@ export const movieCard = ({
           <ul class="button__place">
               <button type="button" class="film-modal__button film-modal__button--active">ADD TO WATCHED</button>
               <button type="button" class="film-modal__button">ADD TO QUEUE</button>
-          </ul>
-
-  </div>`;
+          </ul>`;
 };
